validacao pelo back, mensagens expostas no front
