+++ front/src/app/environments/global-types.d.ts
@@ -1,5 +1,5 @@
 interface Car {
-  id: number;
+  id?: number;
   brand: string;
   model: string;
   color: string;
