+++ front/src/app/modules/car-list/car-list.component.html
@@ -3,7 +3,7 @@
   <p>Aqui está uma lista dos carros disponíveis. Clique em qualquer item para ver mais informações</p>
 </div>
 <div class="d-flex align-items-flex-start justify-content-between">
-  <div class="col-6" *ngIf="cars; else noContent">
+  <div class="col-6" *ngIf="cars && cars.length > 0; else noContent">
     <table class="table table-striped">
     <thead>
         <tr>
@@ -22,7 +22,7 @@
                 <i class="bi bi-pencil-fill"></i>
               </button>
               
-              <button class="btn btn-danger" (click)="delete(car.id)">
+              <button class="btn btn-danger" *ngIf="car.id" (click)="delete(car.id)">
                 <i class="bi bi-trash-fill"></i>
               </button>
             </div>
@@ -59,12 +59,11 @@
     <div class="col-xs">
     <h4>Atualizar dados</h4>
     <p>Atualizando o veículo {{car.brand}} {{car.model}}</p>
-    <app-car-register [id]="car.id"></app-car-register>
+    <car-form *ngIf="car.id" [id]="car.id"></car-form>
     </div>
   </div>
 </ng-template>
 
 <ng-template #noContent>
-  <p>Nenhum veículo disponível!</p> <br/>
-  <p>Tente mais tarde</p>
+  <p>Nenhum veículo disponível!</p>
 </ng-template>
