+++ front/src/app/modules/car-register/car-register.component.html
@@ -1,75 +1,8 @@
-<p>Insira os dados do carro:</p>
+<div>
+  <h3> Cadastro </h3>
+  <p>Insira os dados do carro:</p>
+</div>
 
 <div class="col-6">
-  <form [formGroup]="carForm" (ngSubmit)="onSubmit()" (ngModelChange)="onChange()">
-    <div class="form-group mt-2">
-      <label>Marca</label> <br>
-      <mat-form-field>
-        <mat-select
-          formControlName="brand"
-          (selectionChange)="onBrandChange($event)"
-          placeholder="Selecione uma marca"
-        >
-          <mat-option *ngFor="let brand of brands" [value]="brand">
-            {{ brand != null ? brand.name : "oi" }}
-          </mat-option>
-        </mat-select>
-      </mat-form-field>
-    </div>
-    <div class="form-group mt-2">
-      <label>Modelo</label> <br>
-      <mat-form-field>
-        <mat-select
-          formControlName="model"
-          placeholder="Selecione um modelo"
-          [matTooltip]="carForm.value.brand == null ? 'Você precisa selecionar uma marca primeiro' : null">
-          <mat-option *ngFor="let model of models" [value]="model">
-            {{ model.name }}
-          </mat-option>
-        </mat-select>
-      </mat-form-field>
-    </div>
-    <div class="form-group mt-2">
-      <label>Cor</label>
-      <input
-        class="form-control"
-        type="text"
-        id="color"
-        name="color"
-        formControlName="color"
-        placeholder="Entre uma cor, ex.: azul, amarelo"
-      />
-    </div>
-    <div class="form-group mt-2">
-      <label>Tipo de transmissão</label> <br>
-      <mat-form-field>
-        <mat-select formControlName="transmission">
-          <mat-option *ngFor="let transmissionType of transmissionTypes" [value]="transmissionType">
-            {{ transmissionType.presentableName }}
-          </mat-option>
-        </mat-select>
-      </mat-form-field>
-    </div>
-    <div class="form-group mt-2">
-      <label>Ano de fabricação</label>
-      <input class="form-control" type="text" id="year" name="year" formControlName="year" />
-    </div>
-    <label>Preço do automóvel</label>
-    <div class="input-group">
-      <div class="input-group-prepend">
-        <div class="input-group-text">R$</div>
-      </div>
-      <input class="form-control" type="text" id="price" name="price" formControlName="price" />
-    </div>
-    <ng-container *ngIf="submitted" #submitContainer></ng-container>
-    <div class="form-group">
-      <button [class]="btnType" type="submit" [disabled]="carForm.invalid && !id">Cadastrar</button>
-    </div>
-  </form>
+  <car-form></car-form>
 </div>
-<ng-template #success>
-  <small style="color: green">Dados atualizados com sucesso!</small>
-</ng-template>
-<ng-template #fail>
-  <small style="color: red">Não foi possível atualizar os dados. Tente novamente</small>
-<ng-template>
